requirements.txt
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0
nasdaqdatalink==1.0.4
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.24.0
google-auth==2.25.2
pyarrow==14.0.2
db-dtypes==1.2.0
pydantic==2.5.3
streamlit==1.32.0
plotly==5.19.0
